small product tile + refactor components
