+++ react-app/src/App.js
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from "react";
 import { useDispatch } from 'react-redux'
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import LoginForm from "./components/auth/LoginForm";
-import SignUpForm from "./components/auth/SignUpForm";
-import NavBar from "./components/NavBar";
+import LoginForm from "./components/auth/LoginForm/LoginForm";
+import SignUpForm from "./components/auth/SignUpForm/SignUpForm";
+import NavBar from "./components/NavBar/NavBar";
 import ProtectedRoute from "./components/auth/ProtectedRoute";
-import UsersList from "./components/UsersList";
-import User from "./components/User";
-import HomePage from "./components/HomePage";
+import UsersList from "./components/UsersList/UsersList";
+import User from "./components/User/User";
+import HomePage from "./components/HomePage/HomePage";
 import { authenticate } from "./store/session";
 
 function App() {
