+++ react-app/src/components/NavBar/NavBar.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import LogoutButton from './auth/LogoutButton';
-import Search from './Search';
+import LogoutButton from './../auth/LogoutButton/LogoutButton';
+import Search from '../Search/Search';
+import './NavBar.css'
 
 const NavBar = () => {
   return (
