+++ react-app/src/components/ProductCard/ProductCard.js
@@ -1,0 +1,29 @@
+import React, {useState, useEffect} from 'react'
+import { useDispatch, useSelector } from 'react-redux';
+import {useParams, Link, NavLink} from "react-router-dom"
+import SmallProductTile from '../SmallProductTile/SmallProductTile.js'
+import "./ProductCard.css"
+
+function ProductCard() {
+    // const dispatch = useDispatch()
+    const products = useSelector(state => {
+        return state.product
+    })
+    if(!products) return null
+
+    return (
+        //we may want to limit on the display
+        <div className="product-card-container" style={{display:"flex", flexWrap:'wrap', width:'fit-content'}}>
+            <div className="product-list"style={{display:'flex',flexDirection: 'row'}}>
+                {Object.values(products).map((product,i) => (
+                  <SmallProductTile product={product} key={i} />
+                ))}
+            </div>
+            
+				
+		</div>
+	);
+
+}
+
+export default ProductCard;
