+++ react-app/src/components/SmallProductTile/SmallProductTile.css
@@ -1,0 +1,3 @@
+body{
+    background-color: aliceblue;
+}
