+++ react-app/src/components/SmallProductTile/index.js
@@ -1,0 +1,3 @@
+import SmallProductTile from "./SmallProductTile"
+
+export default SmallProductTile
