+++ react-app/src/components/User/User.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
+import "./User.css"
 
 function User() {
   const [user, setUser] = useState({});
