+++ react-app/src/components/auth/LogoutButton/LogoutButton.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { logout } from "../../store/session";
-import { removeCart } from "../../store/cart"
-import { removeProducts } from "../../store/product";
-
+import { logout } from "../../../store/session";
+import { removeCart } from "../../../store/cart"
+import { removeProducts } from "../../../store/product";
+import './LogoutButton.css'
 const LogoutButton = () => {
   const dispatch = useDispatch()
   const onLogout = async (e) => {
