+++ react-app/src/components/auth/SignUpForm/SignUpForm.js
@@ -1,8 +1,9 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux"
 import { Redirect } from 'react-router-dom';
-import { signUp } from '../../store/session';
-import { createCart } from '../../store/cart';
+import { signUp } from '../../../store/session';
+import { createCart } from '../../../store/cart';
+import './SignUpForm.css';
 
 const SignUpForm = () => {
   const dispatch = useDispatch()
